requirements.txt
requests
pandas
openpyxl
schedule
python-dotenv
